+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Hide burned tokens from collection results in query
+
+## [0.7.3] - 2021-12-08
+- Convert fetch for HEAD to https uris to prevent CORS error
 
 ## [0.8.0] - 2021-12-16
 - Fix ENS resolution to properly use reverse dns, simplify return object